(!) Changed 4 callback names and parameter orders
Container:

Changed the callbacks:
OnPlayerAddToContainer
OnPlayerAddedToContainer
OnPlayerTakeFromContainer
OnPlayerTakenFromContainer

To:
OnItemAddToContainer
OnItemAddedToContainer
OnItemRemoveFromContainer
OnItemRemovedFromContainer

The parameter order for these callbacks has changed from:
(playerid, containerid, itemid)

To:
(containerid, itemid, playerid)

The callbacks are now called whenever the action happens, regardless of
whether a player triggered it or not. the 'playerid' parameter will be
equal to 65535 (INVALID_PLAYER_ID) if the player didn't cause the event.

Item:

Improved the PlayerPickUpItem function to do all the math and variable
setting rather than doing some of the work on OnPlayerKeyStateChange.
This allows better external use of the function.