Major update to the way items link to buttons
Overall:
Added version numbers.

Button:
Changed detection zone from circle back to sphere due to the lack of
height bounds for cylinders in the streamer plugin.

Fixed some bad parameter names.

Inventory:
Added the correct hook for "init" callback with a filterscript check.

Item:
Changed ITEM_BTN_OFFSET_Z name to ITEM_BUTTON_OFFSET, changed its value
to 0.7 and actually used the definition in the script (it was previously
unused even though the value was used in places where the constant
should have been)

Changed the way items lookup button IDs, now there is a global array
where the cell address is a button ID and the value stored in the cell
is an item ID. This makes button lookups a lot faster as there is no
need for a loop, the code simply uses the buttonid from OnButtonPress as
the cell address and retrieves the item ID from the array cell. Much
better with 4k+ items in the game world (See Scavenge and Survive for an
example of this amount of items)

Modules/Notebook:
Added a filterscript check for the OnGameMode/FilterScriptInit hook.