Improved lookups by adding relational indexes
Container:
Added "cnt_ItemContainer" array that contains a cell for each item, this
cell stores the container ID that this item is stored in. This is a much
better alternative to looping through every container to check if the
item is inside it.

Inventory:
Added a similar index to above, each item cell stores a player ID if
that item is stored inside their inventory. As a better alternative to
looping all players and checking if that item is in their inventory.