Another small fix to item giving
Added IsPlayerConnected checks to the GiveItemDelay timer just in case
the receiver quits the game before the GiveItemDelay function could be
called.

Left out a check for the giver, that way if the giver quits, the item
won't be lost in limbo and the receiver should still receive the item
even after the giver quits.