Added debug script, fixed some minor bugs
Button:
Added debug functions to the entire script. Should be improved to output
formatted data however.

Fixed some typos in streamer enumerator labels.

Container:
Improved RemoveItemFromContainer by adding an extra item validity check.
If an item in any of the slots is invalid, it is removed and the list
shifted upwards.

Inventory:
Similar change to RemoveItemFromInventory, if the item in the specified
slot is invalid and the slot still has an ID in it, the slot is emptied
and the list shifted upwards.

Item:
Added ITM_DROP_ON_DEATH compilation control.

Added a world and interior variable to item data. NOTE: The position,
world and interior variables are likely to be removed in the future.
Position, world and interior will be stored only with the button and not
with the item as well.

Added Get/Set functions for item world and interior.

extensions/ContainerDialog:
Fixed a broken ALS hook.

Debug:
Contains functions for runtime debugging. Is not activated at
compiletime but can be turned on/off for either each player or just the
server to log certain script actions.

Plans:
Add a formatted string input method to log specific data such as IDs.
Add script-specific debugging (in case you only want to see debug
messages from SIF/Item for example)