Updated debug module
Drastically improves the usability of the debug functions and
organisation of debugging.

A debug "handler" is registered on script init for each module using
"sif_debug_register_handler". The returned handle ID is then used in the
debug functions. This allows a single script's debug level to be set so
you only see debug information for that script.

An interface function allows searching of debug handlers by name, not a
very good algorithm yet though. Will be improved.