Syncing latest stable version
GetPlayerButtonList now has a "validate" parameter which will ensure the
player is actually near all the buttons. This is only temporary while a
possible bug with the script or streamer plugin is tracked down which
results in the player not removing a button from their nearby-list when
leaving the area.

Added size validation to CreateContainer.

Added a new animation index to IsIdleAnim.

Fixes to ContainerDialog and InventoryDialog to ensure all text fits in
the list.

The OnPlayerDroppedItem callback is now called after the item is removed
from the player (as it should have been).