Updated ContainerDialog and InventoryDialog
Both these scripts have been updated to better support dialog hooking
and external control. The "Close" functions for both now have a "call"
parameter (default false) which, when true, will call the corresponding
"OnPlayerClose..." callback. To prevent recursion if "ClosePlayer..." is
called from within "OnPlayerClose...". In the internal dialog callback
in this script, "call" is set to true so a manual close of the dialog
always calls "OnPlayerClose..." but a script-side call of
"ClosePlayer..." is dependent on the "call" parameter.