Added virtual parameter to CreateItem
Similar to virtual containers, this parameter creates an item as a
virtual item which doesn't exist in the game world. When the item is
created, it doesn't call OnItemCreateInWorld (and, in turn, doesn't call
the RemoveItemFromContainer code if you're using SIF\Container) so this
parameter is useful if you're creating items and adding them to a
container immediately.