Container API overhaul
Containers are now created as virtual all the time. No coordinates,
world or interior parameters on the constructor any more. An optional
'buttonid' parameter instead is used to bind a container to a location
in the world.

Container opening has been moved to the Container module since it
matches the button integration (though this might change in future
anyway).

Organised Container and ContainerDialog variables and functions now to
be more logically modular.