Item update, added hitpoints, cleaned up code
Items now have a hitpoint value which can be used as a damage/usage
value for tools or a trigger for explosives. If the hitpoints of an item
is set to 0 or below via SetItemHitPoints then it is destroyed.

Fixed a load of issues with Item debug labels and improved their
consistency in the Item module.

Removed most of the data resets in DestroyItem.